+++ src/main/java/net/frozenblock/wilderwild/entity/render/model/JellyfishModel.java
@@ -29,6 +29,7 @@ import net.frozenblock.wilderwild.entity.Jellyfish;
 import net.minecraft.client.model.HierarchicalModel;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.model.geom.PartPose;
+import net.minecraft.client.model.geom.builders.CubeDeformation;
 import net.minecraft.client.model.geom.builders.CubeListBuilder;
 import net.minecraft.client.model.geom.builders.LayerDefinition;
 import net.minecraft.client.model.geom.builders.MeshDefinition;
@@ -46,6 +47,7 @@ public class JellyfishModel<T extends Jellyfish> extends HierarchicalModel<T> {
 	private final ModelPart body;
 	private final ModelPart tentacleBase;
 	private final ModelPart[] tentacles = new ModelPart[JELLYFISH_TENTACLES];
+	private final ModelPart[] planeTentacles = new ModelPart[JELLYFISH_TENTACLES];
 	public float xRot;
 	public float tentXRot;
 
@@ -59,7 +61,8 @@ public class JellyfishModel<T extends Jellyfish> extends HierarchicalModel<T> {
 		this.bone = root.getChild("bone");
 		this.body = this.bone.getChild("body");
 		this.tentacleBase = this.bone.getChild("tentacleBase");
-		Arrays.setAll(this.tentacles, i -> tentacleBase.getChild(createTentacleName(i)));
+		Arrays.setAll(this.tentacles, i -> tentacleBase.getChild(createTentacleName(i, false)));
+		Arrays.setAll(this.planeTentacles, i -> tentacleBase.getChild(createTentacleName(i, true)));
 	}
 
 	@NotNull
@@ -69,25 +72,46 @@ public class JellyfishModel<T extends Jellyfish> extends HierarchicalModel<T> {
 
 		PartDefinition bone = partDefinition.addOrReplaceChild("bone", CubeListBuilder.create(), PartPose.ZERO);
 
-		bone.addOrReplaceChild("body", CubeListBuilder.create().texOffs(0, 0).addBox(-4.0F, -2F, -4F, 8F, 5F, 8F)
+		bone.addOrReplaceChild("body", CubeListBuilder.create().texOffs(0, 0).addBox(-4F, -2F, -4F, 8F, 5F, 8F)
 			.texOffs(4, 13).addBox(-3F, -1F, -3F, 6F, 3F, 6F), PartPose.ZERO);
 
 		PartDefinition tentacleBase = bone.addOrReplaceChild("tentacleBase", CubeListBuilder.create(), PartPose.ZERO);
 		makeTentacles(tentacleBase, JELLYFISH_TENTACLES);
+		makePlaneTentacles(tentacleBase, JELLYFISH_TENTACLES);
 
 		return LayerDefinition.create(meshDefinition, 64, 64);
 	}
 
 	private static void makeTentacles(PartDefinition partDefinition, int amount) {
-		CubeListBuilder tentacle = CubeListBuilder.create().texOffs(0, 13).addBox(-0.5F, 0F, 0F, 1F, 10F, 1F);
+		CubeListBuilder tentacle = CubeListBuilder.create().texOffs(0, 13)
+			.addBox(-0.5F, 0F, 0F, 1F, 10F, 1F);
 		for (int i = 0; i < amount; ++i) {
 			float rot = i * Mth.PI * 2F /  amount;
-			partDefinition.addOrReplaceChild(createTentacleName(i), tentacle, PartPose.offsetAndRotation(
+			partDefinition.addOrReplaceChild(createTentacleName(i, false), tentacle, PartPose.offsetAndRotation(
 					(float) Math.cos(rot) * 2.5F,
 					0F,
-                    (float) (Math.sin(rot) * 2.5F),
+                    (float) Math.sin(rot) * 2.5F,
 					0F,
-                    i * Mth.PI * -2F / amount + 1.5707963267948966F,
+                    i * Mth.PI * -2F / amount + Mth.HALF_PI,
+					0F
+				)
+			);
+		}
+	}
+
+	private static void makePlaneTentacles(PartDefinition partDefinition, int amount) {
+		CubeListBuilder tentacle = CubeListBuilder.create().texOffs(0, 14)
+			.addBox(-0.5F, 0F, 1F, 1F, 10F, 0F, new CubeDeformation(0.001F));
+		CubeListBuilder altTentacle = CubeListBuilder.create().texOffs(2, 14)
+			.addBox(-0.5F, 0F, 1F, 1F, 10F, 0F, new CubeDeformation(0.001F));
+		for (int i = 0; i < amount; ++i) {
+			float rot = i * Mth.PI * 2F /  amount;
+			partDefinition.addOrReplaceChild(createTentacleName(i, true), i % 2 == 0 ? altTentacle : tentacle, PartPose.offsetAndRotation(
+					(float) Math.cos(rot) * 2.75F,
+					0F,
+					(float) Math.sin(rot) * 2.75F,
+					0F,
+					i * Mth.PI * -2F / amount + Mth.HALF_PI,
 					0F
 				)
 			);
@@ -95,21 +119,9 @@ public class JellyfishModel<T extends Jellyfish> extends HierarchicalModel<T> {
 	}
 
 	@NotNull
-	private static String createTentacleName(int i) {
-		return "tentacle" + i;
-	}
-
-	private static float fasterRotLerp(float f, float g, float h) {
-		float angleCalc = (h - g) % 360F;
-		if (angleCalc >= 180F) {
-			angleCalc -= 360F;
-		}
-
-		if (angleCalc < -180F) {
-			angleCalc += 360F;
-		}
-
-		return g + f * angleCalc;
+	private static String createTentacleName(int number, boolean plane) {
+		String tentacle = !plane ? "tentacle" : "tentacle_plane";
+		return tentacle + number;
 	}
 
 	@Override
@@ -120,14 +132,15 @@ public class JellyfishModel<T extends Jellyfish> extends HierarchicalModel<T> {
 
 	@Override
 	public void prepareMobModel(@NotNull T jelly, float limbSwing, float limbSwimgAmount, float partialTick) {
-		this.xRot = -(jelly.xRot1 + partialTick * (jelly.xBodyRot - jelly.xRot1));
-		this.tentXRot = -(jelly.xRot6 + partialTick * (jelly.xRot5 - jelly.xRot6));
+		this.xRot = -(jelly.xRot1 + partialTick * (jelly.xBodyRot - jelly.xRot1)) * Mth.DEG_TO_RAD;
+		this.tentXRot = -(jelly.xRot6 + partialTick * (jelly.xRot5 - jelly.xRot6)) * Mth.DEG_TO_RAD;
 	}
 
 	@Override
 	public void setupAnim(@NotNull T jellyfish, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
-		this.bone.xRot = this.xRot * Mth.DEG_TO_RAD;
-		this.tentacleBase.xRot = (this.xRot - this.tentXRot) * Mth.DEG_TO_RAD;
+		this.root().getAllParts().forEach(ModelPart::resetPose);
+		this.bone.xRot = this.xRot;
+		this.tentacleBase.xRot = (this.xRot - this.tentXRot);
 
 		float animation = limbSwing * 2F;
 		float movementDelta = Math.min(limbSwingAmount * 26.6666667F, 1F);
@@ -143,14 +156,25 @@ public class JellyfishModel<T extends Jellyfish> extends HierarchicalModel<T> {
 		this.body.yScale = Mth.lerp(movementDelta, -sinIdle + 1F, 1.25F + (sin * 0.75F));
 
 		this.body.y = Mth.lerp(movementDelta, 0F, 3.5F - (squashStretch * 3.5F));
-		this.tentacleBase.y = Mth.lerp(movementDelta, (-sinIdle * 2F) + 1.8F, (6F - (squashStretch * 5F)) * 1.5F);
+		this.tentacleBase.y = Mth.lerp(movementDelta, (-sinIdle * 2F) + 1.8F, (6F - (squashStretch * 5F)) * 1.5F) * 1.5F;
 
-		float tentRot = -fasterRotLerp(movementDelta, (float) (-Math.sin((ageInTicks - 10) * 0.1F) * 0.2F) + EIGHT_PI, (float) (-Math.sin(animation + 5F) * 20F - 7.5F) * Mth.DEG_TO_RAD);
-		for (ModelPart modelPart : this.tentacles) {
-			PartPose initialPose = modelPart.getInitialPose();
-			modelPart.x = initialPose.x * squash;
-			modelPart.z = initialPose.z * squash;
+		float tentRot = -Mth.rotLerp(
+			movementDelta,
+			(float) (-Math.sin((ageInTicks - 10) * 0.1F) * 0.2F) + EIGHT_PI,
+			(float) (-Math.sin(animation + 5F) * 20F - 7.5F) * Mth.DEG_TO_RAD
+		);
+		ModelPart[] visibleTentacles = !WWEntityConfig.Client.JELLYFISH_PLANE_TENTACLES ? this.tentacles : this.planeTentacles;
+		ModelPart[] invisibleTentacles = !WWEntityConfig.Client.JELLYFISH_PLANE_TENTACLES ? this.planeTentacles : this.tentacles;
+
+		for (ModelPart modelPart : visibleTentacles) {
+			modelPart.visible = true;
+			modelPart.x *= squash;
+			modelPart.z *= squash;
 			modelPart.xRot = tentRot;
+		}
+
+		for (ModelPart modelPart : invisibleTentacles) {
+			modelPart.visible = false;
 		}
 	}